pyproject.toml
[project]
name = "chembot"
version = "1.0.0"
description = "Educational AI Chatbot with RAG"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ChemBot Team"}
]

dependencies = [
    "fastapi==0.109.2",
    "uvicorn[standard]==0.27.1",
    "python-multipart==0.0.9",
    "pymongo==4.6.1",
    "motor==3.3.2",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "bcrypt==4.0.1",
    "python-dotenv==1.0.1",
    "pyjwt==2.8.0",
    "redis==5.0.1",
    "hiredis==2.3.2",
    "pydantic==2.6.1",
    "pydantic-settings==2.1.0",
    "email-validator==2.1.0.post1",
    "PyMuPDF4LLM==0.0.5",
    "pymupdf>=1.24.2",
    "python-docx==1.1.0",
    "pdf2docx==0.5.8",
    "markdown==3.5.2",
    "litellm==1.24.3",
    "pinecone-client==3.0.3",
    "weaviate-client==4.4.1",
    "sentence-transformers==2.3.1",
    "tiktoken==0.6.0",
    "aiofiles==23.2.1",
    "tqdm==4.66.1",
    "PyYAML==6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-asyncio==0.23.3",
    "httpx==0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

